First installment on \x/ill Uther's changes.
While some (like chmod) are clearly Mac motivated, others (like Javac) are
fixes to bugs that can appear on any platform.
Submitted by: William Uther <will+ant@cs.cmu.edu>


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@267621 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/taskdefs/Delete.java
@@ -72,10 +72,14 @@ public class Delete extends Task {
     }
 
     public void execute() throws BuildException {
-        project.log("Deleting: " + f.getAbsolutePath());
-
         if (f.exists()) {
-            f.delete();
+            if (f.isDirectory()) { 
+                project.log("Directory: " + f.getAbsolutePath() + 
+                    " cannot be removed with delete.  Use Deltree instead.");
+            } else {
+                project.log("Deleting: " + f.getAbsolutePath());
+                f.delete();
+            }
         }
     }
 }
